moved border box to new file and renamed it data box
